Use 2d stretch mode to prevent jagged text
More information about the different stretch modes can be found here:
https://docs.godotengine.org/en/stable/tutorials/rendering/multiple_resolutions.html
I think the reason we were seeing badly-rendered text was the viewport
stretch mode, which made the game render in a different resolution and
then get stretched to fit the screen. With 2d mode, it will actually
render at the size it is, which means the font renderer can do
its job of keeping the fonts looking good.